arclink tests: fix cleanup of tmp dcid password file

+++ obspy/clients/arclink/tests/test_decrypt.py
@@ -101,15 +101,19 @@ class ClientTestCase(unittest.TestCase):
         Use $HOME/dcidpasswords.txt.
         """
         dcidfile = DCID_KEY_FILE
-        fh = open(dcidfile, 'wt')
-        fh.write('TEST=XYZ\r\nBIA=OfH9ekhi\r\n')
-        fh.close()
-        # test server for encryption
-        client1 = Client(host="webdc.eu", port=36000, user="test@obspy.org")
-        # public server
-        client2 = Client(host="webdc.eu", port=18001, user="test@obspy.org")
-        # clean up dcid file
-        os.remove(dcidfile)
+        try:
+            with open(dcidfile, 'wt') as fh:
+                fh.write('TEST=XYZ\r\nBIA=OfH9ekhi\r\n')
+            # test server for encryption
+            client1 = Client(host="webdc.eu", port=36000,
+                             user="test@obspy.org")
+            # public server
+            client2 = Client(host="webdc.eu", port=18001,
+                             user="test@obspy.org")
+            # clean up dcid file
+        finally:
+            if os.path.exists(dcidfile):
+                os.remove(dcidfile)
         # request data
         start = UTCDateTime(2010, 1, 1, 10, 0, 0)
         end = start + 100
